pull request para corregir errores de test

+++ src/view-controller.js
@@ -2,18 +2,18 @@ import { checkInFunction, singInFunction, logOut, getUserPostData,
          deletePost, updatePost} from './controller-function/function-firebase.js';
 
 export const checkInOnSubmit = () => {
-  const name = document.querySelector('#user-name').value;
+  const displayName = document.querySelector('#user-name').value;
   const information = document.querySelector('#information').value;
   const email = document.querySelector('#email').value;
   const password = document.querySelector('#password').value;
 
   const firestore = firebase.firestore();
-  let nameUser = name;
+  let nameUser = displayName;
   let informationUser = information;
-  let emailRef = email;
+  let emailRef = email;          
   let users = firestore.collection('users');
   let data = {
-    name: nameUser,
+    displayName: nameUser,
     information: informationUser,
     email: emailRef,
   };
@@ -37,24 +37,27 @@ export const signInOnSubmit = () => {
     });
 };
 
+
+
 export const addPostOnSubmit = () => {
   const contentPost = document.querySelector('#text-area');
   getUserPostData(contentPost.value)
-    .then(() => {})
-    .catch(() => {});
-  // .then((
-  // swal('¡Genial!', 'Tu post se subió satisfactoriamente', 'success'))
-  // .catch((err) => error => {
-  //   console.error('Error adding document: ', error);
-  // })
+  .then(() => {})
+  .catch(() => {})
 };
 
-export const deletePostOnSubmit = (objPost) => 
-  deletePost(objPost.id);
+export const deletePostOnSubmit = (objPost) => { 
+  if(confirm('estas recontra segur de eliminar...'))
+  if(confirm('seguro seguro seguro muy seguro muy pero muy seguro')) 
+  if(confirm('alerta alerta alerta alerta se esta eliminando.........!!!!!!!!!!')){
+   return deletePost(objPost.id);
+  }
+}
 
 export const updatePostSubmit = (objPost, content) => {
   return updatePost(objPost, content);
 };
+
 
 
 export const logOutOnSubmit = () => {
@@ -64,3 +67,4 @@ export const logOutOnSubmit = () => {
     })
     .catch(() => { });
 };
+
